// src/generated/nebula/common.ts
export interface ClientInfo {
  lang: ClientInfo_Language;
  protocolVersion: Buffer;
  version: Buffer;
}

export enum ClientInfo_Language {
  UNKNOWN = 0,
  CPP = 1,
  GO = 2,
  JAVA = 3,
  PYTHON = 4,
  JAVASCRIPT = 5,
}
// ...
export const ClientInfo: MessageFns<ClientInfo> = {
  encode(message: ClientInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.lang !== 0) {
      writer.uint32(8).int32(message.lang);
    }
    if (message.protocolVersion.length !== 0) {
      writer.uint32(18).bytes(message.protocolVersion);
    }
    if (message.version.length !== 0) {
      writer.uint32(26).bytes(message.version);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ClientInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseClientInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.lang = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.protocolVersion = Buffer.from(reader.bytes());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.version = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ClientInfo {
    return {
      lang: isSet(object.lang) ? clientInfo_LanguageFromJSON(object.lang) : 0,
      protocolVersion: isSet(object.protocolVersion)
        ? Buffer.from(bytesFromBase64(object.protocolVersion))
        : Buffer.alloc(0),
      version: isSet(object.version) ? Buffer.from(bytesFromBase64(object.version)) : Buffer.alloc(0),
    };
  },

  toJSON(message: ClientInfo): unknown {
    const obj: any = {};
    if (message.lang !== 0) {
      obj.lang = clientInfo_LanguageToJSON(message.lang);
    }
    if (message.protocolVersion.length !== 0) {
      obj.protocolVersion = base64FromBytes(message.protocolVersion);
    }
    if (message.version.length !== 0) {
      obj.version = base64FromBytes(message.version);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ClientInfo>, I>>(base?: I): ClientInfo {
    return ClientInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ClientInfo>, I>>(object: I): ClientInfo {
    const message = createBaseClientInfo();
    message.lang = object.lang ?? 0;
    message.protocolVersion = object.protocolVersion ?? Buffer.alloc(0);
    message.version = object.version ?? Buffer.alloc(0);
    return message;
  },
};

function bytesFromBase64(b64: string): Uint8Array {
  return Uint8Array.from(globalThis.Buffer.from(b64, "base64"));
}

function base64FromBytes(arr: Uint8Array): string {
  return globalThis.Buffer.from(arr).toString("base64");
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isObject(value: any): boolean {
  return typeof value === "object" && value !== null;
}

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}