// src/generated/nebula/graph.ts
export interface ExecuteRequest {
  sessionId: bigint;
  stmt: Buffer;
}

export interface ExecuteResponse {
  status?: Status | undefined;
  result?: VectorResultTable | undefined;
  summary?: Summary | undefined;
  cursor: Buffer;
}

export interface AuthRequest {
  username: Buffer;
  authInfo: Buffer;
  clientInfo?: ClientInfo | undefined;
}

export interface AuthResponse {
  status?: Status | undefined;
  sessionId: bigint;
  version: Buffer;
}
// ...
export const ExecuteResponse: MessageFns<ExecuteResponse> = {
  encode(message: ExecuteResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      Status.encode(message.status, writer.uint32(10).fork()).join();
    }
    if (message.result !== undefined) {
      VectorResultTable.encode(message.result, writer.uint32(18).fork()).join();
    }
    if (message.summary !== undefined) {
      Summary.encode(message.summary, writer.uint32(26).fork()).join();
    }
    if (message.cursor.length !== 0) {
      writer.uint32(34).bytes(message.cursor);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecuteResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecuteResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = Status.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.result = VectorResultTable.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.summary = Summary.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.cursor = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecuteResponse {
    return {
      status: isSet(object.status) ? Status.fromJSON(object.status) : undefined,
      result: isSet(object.result) ? VectorResultTable.fromJSON(object.result) : undefined,
      summary: isSet(object.summary) ? Summary.fromJSON(object.summary) : undefined,
      cursor: isSet(object.cursor) ? Buffer.from(bytesFromBase64(object.cursor)) : Buffer.alloc(0),
    };
  },

  toJSON(message: ExecuteResponse): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = Status.toJSON(message.status);
    }
    if (message.result !== undefined) {
      obj.result = VectorResultTable.toJSON(message.result);
    }
    if (message.summary !== undefined) {
      obj.summary = Summary.toJSON(message.summary);
    }
    if (message.cursor.length !== 0) {
      obj.cursor = base64FromBytes(message.cursor);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ExecuteResponse>, I>>(base?: I): ExecuteResponse {
    return ExecuteResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ExecuteResponse>, I>>(object: I): ExecuteResponse {
    const message = createBaseExecuteResponse();
    message.status = (object.status !== undefined && object.status !== null)
      ? Status.fromPartial(object.status)
      : undefined;
    message.result = (object.result !== undefined && object.result !== null)
      ? VectorResultTable.fromPartial(object.result)
      : undefined;
    message.summary = (object.summary !== undefined && object.summary !== null)
      ? Summary.fromPartial(object.summary)
      : undefined;
    message.cursor = object.cursor ?? Buffer.alloc(0);
    return message;
  },
};

function createBaseAuthRequest(): AuthRequest {
  return { username: Buffer.alloc(0), authInfo: Buffer.alloc(0), clientInfo: undefined };
}

export const AuthRequest: MessageFns<AuthRequest> = {
  encode(message: AuthRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.username.length !== 0) {
      writer.uint32(10).bytes(message.username);
    }
    if (message.authInfo.length !== 0) {
      writer.uint32(18).bytes(message.authInfo);
    }
    if (message.clientInfo !== undefined) {
      ClientInfo.encode(message.clientInfo, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuthRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuthRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.username = Buffer.from(reader.bytes());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.authInfo = Buffer.from(reader.bytes());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.clientInfo = ClientInfo.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AuthRequest {
    return {
      username: isSet(object.username) ? Buffer.from(bytesFromBase64(object.username)) : Buffer.alloc(0),
      authInfo: isSet(object.authInfo) ? Buffer.from(bytesFromBase64(object.authInfo)) : Buffer.alloc(0),
      clientInfo: isSet(object.clientInfo) ? ClientInfo.fromJSON(object.clientInfo) : undefined,
    };
  },

  toJSON(message: AuthRequest): unknown {
    const obj: any = {};
    if (message.username.length !== 0) {
      obj.username = base64FromBytes(message.username);
    }
    if (message.authInfo.length !== 0) {
      obj.authInfo = base64FromBytes(message.authInfo);
    }
    if (message.clientInfo !== undefined) {
      obj.clientInfo = ClientInfo.toJSON(message.clientInfo);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AuthRequest>, I>>(base?: I): AuthRequest {
    return AuthRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AuthRequest>, I>>(object: I): AuthRequest {
    const message = createBaseAuthRequest();
    message.username = object.username ?? Buffer.alloc(0);
    message.authInfo = object.authInfo ?? Buffer.alloc(0);
    message.clientInfo = (object.clientInfo !== undefined && object.clientInfo !== null)
      ? ClientInfo.fromPartial(object.clientInfo)
      : undefined;
    return message;
  },
};

function createBaseAuthResponse(): AuthResponse {
  return { status: undefined, sessionId: 0n, version: Buffer.alloc(0) };
}

export const AuthResponse: MessageFns<AuthResponse> = {
  encode(message: AuthResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== undefined) {
      Status.encode(message.status, writer.uint32(10).fork()).join();
    }
    if (message.sessionId !== 0n) {
      if (BigInt.asIntN(64, message.sessionId) !== message.sessionId) {
        throw new globalThis.Error("value provided for field message.sessionId of type int64 too large");
      }
      writer.uint32(16).int64(message.sessionId);
    }
    if (message.version.length !== 0) {
      writer.uint32(26).bytes(message.version);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AuthResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAuthResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.status = Status.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.sessionId = reader.int64() as bigint;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.version = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AuthResponse {
    return {
      status: isSet(object.status) ? Status.fromJSON(object.status) : undefined,
      sessionId: isSet(object.sessionId) ? BigInt(object.sessionId) : 0n,
      version: isSet(object.version) ? Buffer.from(bytesFromBase64(object.version)) : Buffer.alloc(0),
    };
  },

  toJSON(message: AuthResponse): unknown {
    const obj: any = {};
    if (message.status !== undefined) {
      obj.status = Status.toJSON(message.status);
    }
    if (message.sessionId !== 0n) {
      obj.sessionId = message.sessionId.toString();
    }
    if (message.version.length !== 0) {
      obj.version = base64FromBytes(message.version);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AuthResponse>, I>>(base?: I): AuthResponse {
    return AuthResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AuthResponse>, I>>(object: I): AuthResponse {
    const message = createBaseAuthResponse();
    message.status = (object.status !== undefined && object.status !== null)
      ? Status.fromPartial(object.status)
      : undefined;
    message.sessionId = object.sessionId ?? 0n;
    message.version = object.version ?? Buffer.alloc(0);
    return message;
  },
};

/** Interface definition */
export type GraphServiceService = typeof GraphServiceService;
// ...
export interface GraphServiceClient extends Client {
  authenticate(
    request: AuthRequest,
    callback: (error: ServiceError | null, response: AuthResponse) => void,
  ): ClientUnaryCall;
  authenticate(
    request: AuthRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: AuthResponse) => void,
  ): ClientUnaryCall;
  authenticate(
    request: AuthRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: AuthResponse) => void,
  ): ClientUnaryCall;
  execute(
    request: ExecuteRequest,
    callback: (error: ServiceError | null, response: ExecuteResponse) => void,
  ): ClientUnaryCall;
  execute(
    request: ExecuteRequest,
    metadata: Metadata,
    callback: (error: ServiceError | null, response: ExecuteResponse) => void,
  ): ClientUnaryCall;
  execute(
    request: ExecuteRequest,
    metadata: Metadata,
    options: Partial<CallOptions>,
    callback: (error: ServiceError | null, response: ExecuteResponse) => void,
  ): ClientUnaryCall;
  streamingExecute(request: ExecuteRequest, options?: Partial<CallOptions>): ClientReadableStream<ExecuteResponse>;
  streamingExecute(
    request: ExecuteRequest,
    metadata?: Metadata,
    options?: Partial<CallOptions>,
  ): ClientReadableStream<ExecuteResponse>;
}

export const GraphServiceClient = makeGenericClientConstructor(
  GraphServiceService,
  "nebula.proto.graph.GraphService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): GraphServiceClient;
  service: typeof GraphServiceService;
  serviceName: string;
};
// ...
type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}